VillageGrpcServiceとしてVillage関連のGrpcServiceを集約
